Moved receipt preview to preview package
